[SOUND EDITOR][REFACTO] : split components and styles in different files and folders
