+++ src/modules/editor/SoundEditor/constants/instruments.ts
@@ -1,0 +1,9 @@
+export const defaultInstruments: Map<string, string> = new Map([
+  ["piano", "Piano"],
+  ["guitar", "Guitar"],
+  ["flute", "Flute"],
+  ["trumpet", "Trumpet"],
+  ["contrabass", "Contrabass"],
+  ["harmonica", "Harmonica"],
+]);
+
